+++ CHANGELOG.md
@@ -18,6 +18,7 @@
   * Added MPTherionRunner ok < warning < error escalation test.
   * Added multi line selection + clickable warnings and errors test.
   * MPTherionRunner checking exit code at end of therion run.
+  * Fixing therion run for flatpak releases.
 
 ## 0.2.36 - 2026-02-24 - The [Danaids](https://en.wikipedia.org/wiki/Danaids) release
 * New features: